+++ CHANGELOG.md
@@ -85,3 +85,4 @@
 ### 1.2.3 (in progress)
 
 - Add currency symbol to the amount field(s)
+- Impose boolean constraint on boolean/integer database fields