+++ CHANGELOG.md
@@ -85,4 +85,5 @@
 ### 1.2.3 (in progress)
 
 - Add currency symbol to the amount field(s)
+- Update form styles to be more browser compatible
 - Impose boolean constraint on boolean/integer database fields